refactor: simplify architecture and fix PostgreSQL configuration
  - Remove unnecessary service interfaces, convert to direct classes
  - Delete *ServiceImpl files, merge functionality into Services
  - Fix Docker PostgreSQL credentials and schema configuration
  - Update Flyway scripts to use 'pagamentos' schema
  - Add PostgreSQL dialect configuration and proper permissions
  - Update roadmap documentation with completed refactoring
