+++ README.md
@@ -27,34 +27,36 @@ Project created for study and aims to be a REST api of a CRUD:
 ---
 
 ### Upgrade Plan
-| Description                                     | Done?                 | Date       |
-|-------------------------------------------------|-----------------------|------------|
-| Upgrade Java version to 17                      | :heavy_check_mark:    | 27/09/2022 |
-| Upgrade Gradle version to 7.5.1                 | :heavy_check_mark:    | 27/09/2022 |
-| Upgrade Swagger 2.9 to OpenApiDoc 3.0           | :heavy_check_mark:    | 28/09/2022 |
-| Include on Controller information of OpenApiDoc | :heavy_check_mark:    | 28/09/2022 |
-| Translate project to English                    | :heavy_check_mark:    | 10/10/2022 |
-| Change Controller Input Objects to DTO/Request  | :heavy_check_mark:    | 11/10/2022 |
-| Create adapter layer                            | :heavy_check_mark:    | 11/10/2022 |
-| Change project name to "ecommerce"              | :heavy_check_mark:    | 11/10/2022 |
-| Create Handler exception controller             | :heavy_check_mark:    | 13/10/2022 |
-| Include @Valid and @Validated to Controllers    | :heavy_check_mark:    | 09/01/2023 |
-| Include flyway on project                       | :heavy_check_mark:    | 23/01/2023 |
-| Upgrade Gradle version to 7.6                   | :heavy_check_mark:    | 23/01/2023 |
-| Upgrade Java version to 19                      | :heavy_check_mark:    | 23/01/2023 |
-| Upgrade Spring version to 3.0.2                 | :heavy_check_mark:    | 23/01/2023 |
-| Create a local Docker for Mysql                 | :heavy_check_mark:    | 26/05/2024 |
-| Create a local Docker for Postgresql            | :heavy_check_mark:    | 26/05/2024 |
-| Create db scripts for flyway                    | :heavy_check_mark:    | 26/05/2024 |
-| Finish unit tests                               | :heavy_check_mark:    | 07/09/2025 |
-| Implement Spring Security                       | :black_square_button: |            |
-| Create cache on RedisDB                         | :black_square_button: |            |
-| Study implementation of the record type         | :black_square_button: |            |
-| TestContainers                                  | :black_square_button: |            |
-| Upgrade Java version to 21                      | :black_square_button: |            |
-| Implement integration with Kafka                | :black_square_button: |            |
-| Create integration with MongoDB                 | :black_square_button: |            |
-| Implement Schedulers for processing             | :black_square_button: |            |
+| Description                                                    | Done?                 | Date       |
+|----------------------------------------------------------------|-----------------------|------------|
+| Upgrade Java version to 17                                     | :heavy_check_mark:    | 27/09/2022 |
+| Upgrade Gradle version to 7.5.1                                | :heavy_check_mark:    | 27/09/2022 |
+| Upgrade Swagger 2.9 to OpenApiDoc 3.0                          | :heavy_check_mark:    | 28/09/2022 |
+| Include on Controller information of OpenApiDoc                | :heavy_check_mark:    | 28/09/2022 |
+| Translate project to English                                   | :heavy_check_mark:    | 10/10/2022 |
+| Change Controller Input Objects to DTO/Request                 | :heavy_check_mark:    | 11/10/2022 |
+| Create adapter layer                                           | :heavy_check_mark:    | 11/10/2022 |
+| Change project name to "ecommerce"                             | :heavy_check_mark:    | 11/10/2022 |
+| Create Handler exception controller                            | :heavy_check_mark:    | 13/10/2022 |
+| Include @Valid and @Validated to Controllers                   | :heavy_check_mark:    | 09/01/2023 |
+| Include flyway on project                                      | :heavy_check_mark:    | 23/01/2023 |
+| Upgrade Gradle version to 7.6                                  | :heavy_check_mark:    | 23/01/2023 |
+| Upgrade Java version to 19                                     | :heavy_check_mark:    | 23/01/2023 |
+| Upgrade Spring version to 3.0.2                                | :heavy_check_mark:    | 23/01/2023 |
+| Create a local Docker for Mysql                                | :heavy_check_mark:    | 26/05/2024 |
+| Create a local Docker for Postgresql                           | :heavy_check_mark:    | 26/05/2024 |
+| Create db scripts for flyway                                   | :heavy_check_mark:    | 26/05/2024 |
+| Finish unit tests                                              | :heavy_check_mark:    | 07/09/2025 |
+| Refactor to simplify and adjusts Docker                        | :heavy_check_mark:    | 13/09/2025 |
+| Implement Spring Security                                      | :black_square_button: |            |
+| Create cache on RedisDB                                        | :black_square_button: |            |
+| Study implementation of the record type                        | :black_square_button: |            |
+| Upgrade Java version to 21                                     | :black_square_button: |            |
+| Implement integration with Kafka                               | :black_square_button: |            |
+| Create integration with MongoDB                                | :black_square_button: |            |
+| TestContainers to MongoDB, PostgresSQL, Mysql, Kafka and Redis | :black_square_button: |            |
+| Implement Schedulers for processing                            | :black_square_button: |            |
+| Refactoring to modular monolith                                | :black_square_button: |            |
 
 ---
