+++ src/main/java/br/com/ecommerce/config/PostgresConfig.java
@@ -14,6 +14,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 import javax.sql.DataSource;
+import java.util.Map;
 
 @Configuration
 @EnableTransactionManagement
@@ -38,6 +39,7 @@ public class PostgresConfig {
                 .persistenceUnit("customer")
                 .persistenceUnit("inventory")
                 .persistenceUnit("product")
+                .properties(Map.of("hibernate.dialect", "org.hibernate.dialect.PostgreSQLDialect"))
                 .build();
     }
