+++ src/main/java/br/com/ecommerce/domain/service/ProductService.java
@@ -1,19 +1,68 @@
 package br.com.ecommerce.domain.service;
 
+import br.com.ecommerce.adapter.toentity.ProductRequestToProductEntityAdapter;
+import br.com.ecommerce.adapter.toresponse.ProductEntityToResponseAdapter;
+import br.com.ecommerce.controller.common.enumerated.MessageEnum;
 import br.com.ecommerce.controller.request.ProductRequest;
 import br.com.ecommerce.controller.response.ProductResponse;
+import br.com.ecommerce.domain.entity.postgres.ProductEntity;
+import br.com.ecommerce.domain.repository.postgres.ProductRepository;
 import br.com.ecommerce.infra.exception.ProductCreateException;
 import br.com.ecommerce.infra.exception.ProductDeleteException;
 import br.com.ecommerce.infra.exception.ProductNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface ProductService {
-    ProductResponse retrieveProduct(String id) throws ProductNotFoundException;
+@Service
+public class ProductService {
 
-    List<ProductResponse> retrieveListProducts() throws ProductNotFoundException;
+    private final ProductRepository productRepository;
+    private final ProductRequestToProductEntityAdapter productRequestToProductEntityAdapter;
+    private final ProductEntityToResponseAdapter productEntityToResponseAdapter;
 
-    ProductResponse createProduct(ProductRequest productRequest) throws ProductCreateException;
+    @Autowired
+    public ProductService(ProductRepository productRepository, ProductRequestToProductEntityAdapter productRequestToProductEntityAdapter, ProductEntityToResponseAdapter productEntityToResponseAdapter) {
+        this.productRepository = productRepository;
+        this.productRequestToProductEntityAdapter = productRequestToProductEntityAdapter;
+        this.productEntityToResponseAdapter = productEntityToResponseAdapter;
+    }
 
-    void deleteProduct(String id) throws ProductDeleteException;
+    public ProductResponse retrieveProduct(String id) throws ProductNotFoundException {
+        ProductEntity productEntity =
+                productRepository.findById(id)
+                        .orElseThrow(() -> new ProductNotFoundException(MessageEnum.PRODUCT_NOT_FOUND_EXCEPTION.getValue(), new Exception()));
+
+        return productEntityToResponseAdapter.getProductResponse(productEntity);
+    }
+
+    public List<ProductResponse> retrieveListProducts() throws ProductNotFoundException {
+        List<ProductEntity> productEntityList = this.productRepository.findAll();
+        if (productEntityList.isEmpty())
+            throw new ProductNotFoundException(MessageEnum.PRODUCT_LIST_NOT_FOUND_EXCEPTION.getValue(), new Exception());
+
+        return productEntityToResponseAdapter.buildListProductResponse(productEntityList);
+    }
+
+    @Transactional
+    public ProductResponse createProduct(ProductRequest productRequest) throws ProductCreateException {
+        ProductEntity productEntity = productRequestToProductEntityAdapter.getProductEntity(productRequest);
+        try {
+            ProductEntity productEntitySaved = this.productRepository.save(productEntity);
+            return productEntityToResponseAdapter.getProductResponse(productEntitySaved);
+        } catch (Exception ex) {
+            throw new ProductCreateException(MessageEnum.PRODUCT_ERROR_ON_CREATE_EXCEPTION.getValue(), ex);
+        }
+    }
+
+    @Transactional
+    public void deleteProduct(String id) throws ProductDeleteException {
+        try {
+            this.productRepository.deleteById(id);
+        } catch (Exception ex) {
+            throw new ProductDeleteException(MessageEnum.PRODUCT_ERROR_ON_DELETE_EXCEPTION.getValue(), ex);
+        }
+    }
 }
