+++ src/main/resources/ValidationMessages.properties
@@ -2,4 +2,3 @@ product.code.format=Invalid code format.
 product.category.required=Category is required.
 product.weight.minimum=Minimum value is "0.01".
 product.weight.format=Weight format is NNNNN.NN.
-product.date.format=Date format is yyyy-MM-dd and cannot be a future date.
