+++ src/main/resources/db/postgresql/flyway/V2__Insert_inventory.sql
@@ -1,19 +1,19 @@
-INSERT INTO public.inventory
+INSERT INTO pagamentos.inventory
 (id, availablequantity, productcode, reservedquantity)
 VALUES(1, 100, 'ZZ6781', 5);
 
-INSERT INTO public.inventory
+INSERT INTO pagamentos.inventory
 (id, availablequantity, productcode, reservedquantity)
 VALUES(2, 25, 'JZ7201', 0);
 
-INSERT INTO public.inventory
+INSERT INTO pagamentos.inventory
 (id, availablequantity, productcode, reservedquantity)
 VALUES(3, 290, 'CQZ729', 0);
 
-INSERT INTO public.inventory
+INSERT INTO pagamentos.inventory
 (id, availablequantity, productcode, reservedquantity)
 VALUES(4, 0, 'BY6732', 0);
 
-INSERT INTO public.inventory
+INSERT INTO pagamentos.inventory
 (id, availablequantity, productcode, reservedquantity)
 VALUES(5, 178, 'AA4730', 0);
