+++ src/main/resources/db/postgresql/flyway/V3__Insert_product.sql
@@ -1,15 +1,15 @@
-INSERT INTO public.product
+INSERT INTO pagamentos.product
 (code, category, dateregister, description, title, weight)
 VALUES('ZZ6781', 'tecnologia', '2023-01-25', 'Gabinete', 'Led3D', 2.50);
-INSERT INTO public.product
+INSERT INTO pagamentos.product
 (code, category, dateregister, description, title, weight)
 VALUES('JZ7201', 'tecnologia', '2023-01-24', 'Monitor Gamer 165hz', 'Monitor Gamer 3x', 3.50);
-INSERT INTO public.product
+INSERT INTO pagamentos.product
 (code, category, dateregister, description, title, weight)
 VALUES('CQZ729', 'tecnologia', '2022-12-20', 'Mouse de bolinha', 'Mouse old scholl', 0.30);
-INSERT INTO public.product
+INSERT INTO pagamentos.product
 (code, category, dateregister, description, title, weight)
 VALUES('BY6732', 'tecnologia', '2023-01-26', 'Gabinete', 'Optimus', 3.80);
-INSERT INTO public.product
+INSERT INTO pagamentos.product
 (code, category, dateregister, description, title, weight)
 VALUES('AA4730', 'tecnologia', '2022-03-26', 'Led piscante', 'Led Colorido', 1.15);
