+++ src/main/resources/db/postgresql/flyway/V4__Insert_customer.sql
@@ -1,15 +1,15 @@
-INSERT INTO public.customer
+INSERT INTO pagamentos.customer
 (cpf, birthdate, gender, "name")
 VALUES('12345687912', '2001-01-25', 'M', 'BRUNO');
 
-INSERT INTO public.customer
+INSERT INTO pagamentos.customer
 (cpf, birthdate, gender, "name")
 VALUES('12999687912', '1994-12-10', 'F', 'ELA');
 
-INSERT INTO public.customer
+INSERT INTO pagamentos.customer
 (cpf, birthdate, gender, "name")
 VALUES('55545687912', '1985-06-07', 'M', 'ELE');
 
-INSERT INTO public.customer
+INSERT INTO pagamentos.customer
 (cpf, birthdate, gender, "name")
 VALUES('12345684442', '1963-04-03', 'M', 'OUTRA PESSOA');
