+++ src/test/java/br/com/ecommerce/adapter/toentity/ProductRequestToProductEntityAdapterTest.java
@@ -10,7 +10,6 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.*;
